revise_tes_maths_nuxt/backend/requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==4.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
